% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_confidence_envelop.R
\name{generalCE}
\alias{generalCE}
\title{Compute the confidence envelop for any function of the true null hypotheses}
\usage{
generalCE(pvals, func, alpha = 0.05, nullTest = stoufferTest)
}
\arguments{
\item{pvals}{A named numeric vector of p-values for the testing of the null hypotheses}

\item{func}{A function that takes a vector of indices of the true null hypotheses and returns a numeric value}

\item{alpha}{The significance level}

\item{nullTest}{A function that takes a vector of p-values and alpha level and returns significance of the global test}
}
\description{
Compute the confidence envelop for any function of the true null hypotheses
}
